+++ R/lassosum2.R
@@ -26,10 +26,12 @@ snp_lassosum2 <- function(corr, df_beta,
                           delta = c(0.001, 0.01, 0.1, 1),
                           nlambda = 30, lambda.min.ratio = 0.01,
                           dfmax = 200e3, maxiter = 1000, tol = 1e-5,
+                          ind.corr = cols_along(corr),
                           ncores = 1) {
 
   assert_df_with_names(df_beta, c("beta", "beta_se", "n_eff"))
-  assert_lengths(rows_along(corr), cols_along(corr), rows_along(df_beta))
+  assert_lengths(ind.corr, rows_along(df_beta))
+  stopifnot(all(ind.corr %in% cols_along(corr)))
   assert_pos(df_beta$beta_se, strict = TRUE)
   assert_pos(delta, strict = TRUE)
   assert_cores(ncores)
@@ -58,6 +60,7 @@ snp_lassosum2 <- function(corr, df_beta,
         beta_hat       = beta_hat,
         lambda         = pf * grid_param$lambda[ic],
         delta_plus_one = pf * grid_param$delta[ic] + 1,
+        ind_sub        = ind.corr - 1L,
         dfmax          = dfmax,
         maxiter        = maxiter,
         tol            = tol